# perf(haste-map): default to node crawl if watchman is not enabled