… e2e/__tests__/__snapshots__/showConfig.test.ts.snap