<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. The two fields below are mandatory. -->

<!-- Please remember to update CHANGELOG.md at the root of the project if you have not done so. -->

## Summary

Toggling option from #11264.

On my machine the node crawl is even faster than watchman for `--list-tests`, while both beat `find` into dust.

This branch:

```sh-session
$  hyperfine 'yarn jest --list-tests' 'yarn jest --list-tests --no-watchman'
Benchmark 1: yarn jest --list-tests
  Time (mean ± σ):      2.921 s ±  0.750 s    [User: 3.143 s, System: 0.549 s]
  Range (min … max):    2.261 s …  4.519 s    10 runs

Benchmark 2: yarn jest --list-tests --no-watchman
  Time (mean ± σ):      2.197 s ±  0.216 s    [User: 2.367 s, System: 0.414 s]
  Range (min … max):    2.074 s …  2.801 s    10 runs
```

`main`:

```sh-session
$ hyperfine 'yarn jest --list-tests --no-watchman'
Benchmark 1: yarn jest --list-tests --no-watchman
  Time (mean ± σ):      3.684 s ±  0.364 s    [User: 2.878 s, System: 1.415 s]
  Range (min … max):    3.225 s …  4.341 s    10 runs
```

<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->

## Test plan

Green CI

<!-- Demonstrate the code is solid. Example: The exact commands you ran and their output, screenshots / videos if the pull request changes UI. -->